Export function at index
